Real-space US:
Fixing a problem in mg_eig_state.c. It was using the eigenvalues from 
subspace diagonalization, but in first step, the eigenvalues are undefined since 
subspace diagonalization was not run yet. The fix is for the first step to use
eigenvalues calculated by evaluating <psi|H|psi> and use eigenvalues from subspace 
diagonalization in later steps. This only applies if subspace diagonalization is used in every step.

This should also take care of negative charge we see in the initial steps.


git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@1274 eb38b3b0-33ea-0310-94ad-d43fea0d214f