Real-space US: Modifying Makefiles, so that libxc is compiled and installed during compile of our code. Tested on linux and xt targets.
The way this works is that our make target (rmg) now depends on compiled libxc library
located ../Common/libxc/lib/libxc.a. If this file does not exist, we call Make.rmg in 
libxc directory, which runs configure scripts and the compile libxc and installs the 
library it its expected location. If libxc.a exists, no libxc compilation will take place. 

Finally, Make.rmg in the libxc directory needs to be improved so that libxc gets recompiled
when we upgrade libxc source. Given the late hour, I will have to take care of it later.


git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@1300 eb38b3b0-33ea-0310-94ad-d43fea0d214f