Real-space US: Changing the way how mixed dot products between beta and wave functions (oldsint) are handled in the code. Previously the mixing was done in app_nl_eig, but only when an appropriate flag was passed. I have added a function Common/mix_betaxpsi.c, which takes care of mixing and is called before app_nl_eig. The advantage is that mixing can now be done on the entire array, while before it was done multiple times on smaller parts of array, which is less efficient.
Also, app_nl_eig now takes an sint (either newsint or oldsint) as an argument, which comes in handy since
we sometimes need to call it with mixed and other time with non-mixed coefficients. This change will allow for a 
proper mixing, since the old code, which was hard-coded to used mixed coefficients, would just overwrite 
the mixed data (oldsintR) with unmixed ones (newsintR) whenever it needed to use the unmixed coefficients.
Obviously, to those who got here and actually understand what is going on, this results in the loss of history
for <beta|psi>.



git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@1312 eb38b3b0-33ea-0310-94ad-d43fea0d214f