Adjustments to the build process to make the shared directories work properly. These are now setup in the top level Makefile in the GLOBAL_MODULES (as opposed to LOCAL_MODULES for non-shared directories) variable. The GLOBAL_MODULES does not need to be set up anywhere else. Also, each build process creates a log file named .build.log in code's own directory (i.e. RMG, ON, NEGF).
Finally, to prevent collision in shared objects, which can happen if shared object built for one coed are used in compile
of another, top level Makefile has the ability to remove shared objects if they belong to a different code. Before any of
the coded is compiled, a shell script is called, which compares timestamps on the shared objects against that of the build 
log file. The shared objects that are newer and therefore presumably created by a compile of a different code will be 
removed, which will force their recompile. 


git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@1443 eb38b3b0-33ea-0310-94ad-d43fea0d214f