RMG: Rewrite of betaxpsi1, which calculates <beta|psi> coefficients. Previously, each core would calculate its part of the coefficients and then global array containing all coefficients was summed over all cores using global_sums. For large systems, this requires a large memory and also has substantial communication cost.
New betaxpsi1 only involves communication about ions that only have overlap with a core under consideration. 
Here is how it works: Each ion has a core that "owns" it, this core receives contributions from all other 
cores that have overlap with this ion and sums them up. After this, the owning core sends results to the cores 
that have overlap with this ion. The code uses non-blocking point-to-point communication. For efficiency, all
data to be exchanged between a pair of cores is packed and sent in one call. 

This rewrite is important for scaling to very large systems, the old version has a problem there, since it uses
a lot of memory and requires a lot communication. The new version alleviates both issues. Tests show that the new 
version is much faster, but since calculation of <beta|psi> only becomes important for
truly large systems (~1,000 atoms) the speedup for medium size systems is very small. Also, this change should not
affect results in any way.


git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@1503 eb38b3b0-33ea-0310-94ad-d43fea0d214f