Adjustment to read_pseudo to make sure that filtering information for beta functions is read properly.
The problem we have with current PP format is that for US the lines giving l number for potentials
are useless, since each beta function has its l number defined later. The first set is only used for reading
l-dependent filtering information (nlrcut). The problem is that for several our PPs the two sets of l do not match
- first set specifies more l values - but that in that case it works out, we just end up with more filtering 
information for non-existent beta channel. The code I added protects against opposite case, when we have beta function
for which nlrcut has not been defined.

Going forward, we should modify the current PP format to get rind of this duplicity and also to get 
a rid of fields which are not used at all ( l-value for state which is local).



git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@1627 eb38b3b0-33ea-0310-94ad-d43fea0d214f