Initial commit of LCAO start for RMG calculations. This is still under development and currently only works for atoms. More work is needed on a heuristics for systems with  arbitrary number of unoccupied states. Also, this feature needs new pseudopotentials, the ones that we have in the repository do not contain all the necessary information (atomic charge, occupation of states, etc.). Initial tests for atoms show that LCAO start can decrease the number of SCF steps from 55 (random) to 35. This is still not ideal, but initial eigenvalues and RMS (10^-5) are promising and we should be able to converge even faster than in 35 steps.
The atomic wavefunctions and charge are filtered just like projectors and are evaluated on the fine linear grid and
stored in SPECIES structures in awave_lig and arho_lig. The filtering parameters can be set in PP file on the line
where local potential radius is. The parameters are optional, if none are specified, default values will be used. 

If anyone wants to test it now, you need a new PP (should be in repository soon) and set start_mode to "LCAO Start". 
For now, the number of wavefunctions in the calculations has to be equal to that from PP. 


git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@1635 eb38b3b0-33ea-0310-94ad-d43fea0d214f