Initial commit of mask function filtering. This should be an improvement over what we have been using so far, mainly because the filtered functions are exactly zero at their radii (as defined in PP file) so we do not need to dampen with gaussians. More info in JCP 124, 174102 (2006).
This option is in the process of being tested and is OFF by default. It can be turned on by setting mask_function_filtering
to "true" in the control file. This presently only works fully in the real-space code, for ON it for now only affects
Q function filtering. 

Also, filtering has been abstracted into filter_potential function, which takes care of all filtering (applying mask function, calling
FF, backing out the mask function, applying gaussian dampening, etc.). This should be better that calling all the steps explicitly
each time filtering is applied.



git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@1641 eb38b3b0-33ea-0310-94ad-d43fea0d214f