RMG:
changing default wave file output name, with directory creation there is no
longer a reason to default wave files to the current working directory instead
of a subdirectory.

Adding Dimer relax as an option, doesn't work yet but code is in development.

NEB now defaults to a climbing image with tangents defined by high side energy.

move_ions got indented no other changes.

use local copy of force vector to calculate fast relax updates. This is a bit of a hack as it would be better if global force structure could be used directly.

Tested for compilation and run time on the XK6 partition of Jaguar.



git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@1650 eb38b3b0-33ea-0310-94ad-d43fea0d214f