Fix for compiling new linxc on machines with pgcc. This compiler breaks libxc autoconfig script, the issue is line 162 in config.guess, which is:
eval `$CC_FOR_BUILD -E $dummy.c 2>/dev/null | grep '^LIBC'`

For pgcc compiler the expression in the eval statement evaluates to "LIBC = gnu", which shell cannot 
properly parse because of spaces around the equal sign. As the result, LIBC ends up undefined and the 
build script breaks. My fix tests if $LIBC is empty and if it is sets it to "gnu". 

This is of course, just a hack, proper solution would be to clean the spaces, but I do not want to 
bother with that now.



git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@1662 eb38b3b0-33ea-0310-94ad-d43fea0d214f