fix: convergence plot in RMG/Common/Quench.cpp
Given that the system has PLplot installed and it is to be used in this
program, a compile time error would be shown pointing to the second
argument of the PlotConvergence function, declared as CONVERGED. In the
definition, this value was named `ct.scf_converged` instead, which is an
invalid name for a function argument. Since in the only invokation of
this function ct.scf_converged is passed, the second parameter has been
removed.