fix: missing cuda include directories
The CUDA_INCLUDE_DIRS variable was unset when it was being used as a
parmaeter in the include_directories function. Logic has been added to
make it fallback to a default path. If CUDA 13+ is being used, an
additional include is added; appending /cccl to the end of the original
directory. In this way, CUDA_INCLUDE_DIRS is now treated more as a root
rather than a list of directories.